Add Spanish translation
This is a first attempt of translation to Spanish. I named it to `es_AR` b/c I'm from Argentina and I'm not really sure it's dialect neutral enough. I used the Portuguese translation as a basis for which words to translate. I let the _commit_ word in English b/c the literal translation sounds really weird in Spanish (I think it's something like _confirmación_ according the Pro-Git Spanish translation). The same goes for _merge_, _rebase_, _squash_, which are terms from the GIT domain and I don't think it's worth to translate.