Add note about avoiding personal language
It's best to use impersonal language when writing commits.

It doesn't matter if you or someone else did the work (that's why we
often use Key: Value tags [2] on the trailer region such as
"Reviewed-by: " or "Acked-by: "). Commits have the sole purpose of
proposing changes to code, not to tell __your__ story.

[1] https://subjectguides.york.ac.uk/academic-language/objective
[2] https://blog.elastocloud.org/2016/12/adding-reviewed-by-and-acked-by-tags.html

Signed-off-by: Anne Macedo <annie@retpolanne.com>